ex6
need to fix loader + add dashboard.

Loader and better error handling will be done on monday. Dashboards+ filters on tuesday
