+++ main_exercise/client/public/favicon.ico
