+++ main_exercise/client/public/manifest.json
