+++ main_exercise/client/src/Components/Checkbox/Checkbox.js
@@ -1,5 +1,6 @@
 import {Checkbox} from 'monday-ui-react-core';
 import styles from './Checkbox.module.css';
+import PropTypes from "prop-types";
 const CheckboxComponent = ({onChangeHandler}) => {
     return (
         <Checkbox
@@ -7,6 +8,10 @@ const CheckboxComponent = ({onChangeHandler}) => {
             onChange={onChangeHandler}
         />
     );
+}
+
+CheckboxComponent.propTypes = {
+    onChangeHandler: PropTypes.func.isRequired,
 }
 
 export default  CheckboxComponent
