+++ main_exercise/client/src/Components/Dropdown/Dropdown.js
@@ -3,8 +3,12 @@ import "monday-ui-react-core/dist/main.css";
 import PropTypes from "prop-types";
 import styles from "./Dropdown.module.css";
 
-const DropdownComponent = ({ options,label, selectedValue, onChangeHandler }) => {
-
+const DropdownComponent = ({
+  options,
+  label,
+  selectedValue,
+  onChangeHandler,
+}) => {
   return (
     <Dropdown
       options={options}
@@ -21,6 +25,8 @@ const DropdownComponent = ({ options,label, selectedValue, onChangeHandler }) =>
 DropdownComponent.propTypes = {
   options: PropTypes.arrayOf(PropTypes.object).isRequired,
   label: PropTypes.string.isRequired,
+  selectedValue: PropTypes.array.isRequired,
+  onChangeHandler: PropTypes.func.isRequired,
 };
 
 export default DropdownComponent;
