+++ main_exercise/package.json
@@ -1,22 +1,21 @@
 {
-  "name": "monday-u-exercises-1",
-  "type": "commonjs",
+  "name": "ex6",
   "version": "1.0.0",
+  "description": "",
+  "main": "server.js",
   "scripts": {
-    "start": "node server.js",
-    "start:dev": "nodemon server.js"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start:dev": "nodemon server.js",
+    "start": "node server.js"
   },
+  "author": "",
+  "license": "ISC",
   "dependencies": {
-    "axios": "^0.27.2",
-    "body-parser": "^1.20.0",
-    "express": "4.18.1",
-    "fs": "^0.0.1-security",
+    "express": "^4.18.1",
     "mysql2": "^2.3.3",
-    "node-fetch": "^3.2.6",
-    "sequelize": "^6.21.0"
+    "sequelize": "^6.21.2"
   },
   "devDependencies": {
-    "nodemon": "^1.18.4",
-    "sequelize-cli": "^6.4.1"
+    "nodemon": "^2.0.18"
   }
 }
