+++ main_exercise/server/db/seeders/20220622200337-demo-todo.js
@@ -16,6 +16,7 @@ module.exports = {
       isPokemon: false,
       pokemonID: null,
       status: null,
+      urgency: null,
       createdAt: new Date(),
       updatedAt: new Date()
     }]);
