--- src/ex6/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "ex6",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start:dev": "nodemon server.js",
-    "start": "node server.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "express": "^4.18.1",
-    "mysql2": "^2.3.3",
-    "sequelize": "^6.21.2"
-  },
-  "devDependencies": {
-    "nodemon": "^2.0.18"
-  }
-}
