--- src/ex6/server/db/config/config.json
@@ -1,24 +1,0 @@
-
-{
-  "development": {
-    "username": "root",
-    "password": "password",
-    "database": "tododb",
-    "host": "127.0.0.1",
-    "dialect": "mysql"
-  },
-  "test": {
-    "username": "root",
-    "password": null,
-    "database": "database_test",
-    "host": "127.0.0.1",
-    "dialect": "mysql"
-  },
-  "production": {
-    "username": "root",
-    "password": null,
-    "database": "database_production",
-    "host": "127.0.0.1",
-    "dialect": "mysql"
-  }
-}
