--- src/ex6/server/db/models/todo.js
@@ -1,27 +1,0 @@
-'use strict';
-// import {Model} from "sequelize";
-const {Model} = require('sequelize');
-
-module.exports = (sequelize, DataTypes) => {
-  class Todo extends Model {
-    /**
-     * Helper method for defining associations.
-     * This method is not a part of Sequelize lifecycle.
-     * The `models/index` file will call this method automatically.
-     */
-    static associate(models) {
-      // define association here
-    }
-  }
-  Todo.init({
-    todo: DataTypes.STRING,
-    isPokemon: DataTypes.BOOLEAN,
-    pokemonID: DataTypes.INTEGER,
-    status: DataTypes.STRING,
-    urgency: DataTypes.STRING
-  }, {
-    sequelize,
-    modelName: 'Todo',
-  });
-  return Todo;
-};
