--- src/ex6/server/db/seeders/20220622200337-demo-todo.js
@@ -1,34 +1,0 @@
-'use strict';
-
-module.exports = {
-  async up (queryInterface, Sequelize) {
-    /**
-     * Add seed commands here.
-     *
-     * Example:
-     * await queryInterface.bulkInsert('People', [{
-     *   name: 'John Doe',
-     *   isBetaMember: false
-     * }], {});
-    */
-    return queryInterface.bulkInsert('Todos', [{
-      todo: 'Learn Sequelize',
-      isPokemon: false,
-      pokemonID: null,
-      status: null,
-      urgency: null,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    }]);
-  },
-
-  async down (queryInterface, Sequelize) {
-    /**
-     * Add commands to revert seed here.
-     *
-     * Example:
-     * await queryInterface.bulkDelete('People', null, {});
-     */
-    return queryInterface.bulkDelete('Todos', null, {});
-  }
-};
