--- src/ex6/server/services/item_manager.js
@@ -1,126 +1,0 @@
-// The ItemManager should go here. Remember that you have to export it.
-
-const { Todo } = require("../db/models");
-
-// import PokemonClient from "../clients/pokemon_client.js";
-const PokemonClient = require("../clients/pokemon_client.js");
-const { writeFileSync, readFileSync } = require("fs");
-
-class ItemManager {
-  // Constructor
-  constructor() {
-    this.history = [];
-    this.pokemonClient = new PokemonClient();
-    this.todoList = this.getTodoList() || [];
-  }
-
-  trimSpaces(splitList) {
-    const regex = /\s/g;
-    let cleanList = splitList.map((element) => element.replace(regex, ""));
-    return cleanList;
-  }
-
-  checkNumbers(splitList) {
-    const regex = /^\d+$/;
-    return this.trimSpaces(splitList).every((element) => regex.test(element));
-  }
-
-  async addTodo(todo) {
-    const splitList = todo.split(",");
-    const renderedTodos = [];
-    if (this.checkNumbers(splitList)) {
-      const pokemonIDArray = splitList;
-      const pokemonNameArray = await this.pokemonClient.getPokemons(
-        this.trimSpaces(splitList)
-      );
-      for (const element of pokemonNameArray) {
-        const todoObject = {
-          todo: element,
-          pokemonID: pokemonIDArray[pokemonNameArray.indexOf(element)],
-          isPokemon: true,
-        };
-        const {
-            dataValues: { id },
-        } = await Todo.create({
-          todo: todoObject.todo,
-          pokemonID: todoObject.pokemonID,
-          isPokemon: todoObject.isPokemon,
-          status: todoObject.status,
-          urgency: todoObject.urgency
-        });
-        renderedTodos.push({ ...todoObject, id });
-      }
-    } else {
-      const todoObject = {
-        todo: todo,
-        isPokemon: false,
-      };
-      const {dataValues:{id}} = await Todo.create({
-        todo: todoObject.todo,
-        pokemonID: todoObject.pokemonID,
-        isPokemon: todoObject.isPokemon,
-        status: todoObject.status,
-        urgency: todoObject.urgency
-      });
-      renderedTodos.push({ ...todoObject, id });
-    }
-    return renderedTodos;
-  }
-
-  async getTodoList() {
-    try {
-      const data = (await Todo.findAll()).map((element) =>
-        element.get({ plain: true })
-      );
-      return data;
-    } catch {
-      console.log("No todo list found");
-      return [];
-    }
-  }
-
-  async deleteTodo(id) {
-    await Todo.destroy({
-      where: { id },
-    });
-  }
-  async deleteSelected(idArray) {
-    idArray.forEach((element) => {
-      try{
-        Todo.destroy({
-          where: {id: element},
-        });
-      }
-      catch (e) {
-        console.log(e)
-      }})};
-
-  async clearTodoList() {
-    await Todo.destroy({
-      where: {},
-    });
-  }
-
-  async updateStatus(id, status) {
-    await Todo.update(
-      {
-        status: status,
-      },
-      {
-        where: { id },
-      }
-    );
-  }
-  async updateUrgency(id, urgency){
-    await Todo.update(
-        {
-          urgency: urgency,
-        },
-        {
-          where: {id},
-        }
-    );
-  }
-}
-
-module.exports = ItemManager;
